Fix render commands batching on Emscripten
- Add a configurable option for Emscripten fixed batch size
- Fix max batch size count in the render batcher class
- Stabilize render commands sorting with an id based secondary key
- Fix support for unsigned integers in shaders
- Change mesh index shader variable type to unsigned
- Avoid creating a log file if the log level is off
- Add FULL_ES2=1 Emscripten compiler definition to fix error with glVertexAttribIPointer
- Fix OpenGL ES version detection on Emscripten
- Fix some minor signedness and type conversion warnings

+++ src/graphics/RenderBuffersManager.cpp
@@ -40,7 +40,7 @@ RenderBuffersManager::RenderBuffersManager(bool useBufferMapping, unsigned long 
 	uboSpecs.target = GL_UNIFORM_BUFFER;
 	uboSpecs.mapFlags = useBufferMapping ? GL_MAP_WRITE_BIT | GL_MAP_INVALIDATE_BUFFER_BIT | GL_MAP_FLUSH_EXPLICIT_BIT : 0;
 	uboSpecs.usageFlags = GL_STREAM_DRAW;
-	uboSpecs.maxSize = static_cast<unsigned int>(uboMaxSize);
+	uboSpecs.maxSize = static_cast<unsigned long>(uboMaxSize);
 	uboSpecs.alignment = static_cast<unsigned int>(offsetAlignment);
 
 	// Create the first buffer for each type right away
